requirements.txt
python-telegram-bot==20.3
requests>=2.28.0
aiohttp>=3.8.1
pyyaml>=6.0
gitpython>=3.1.30
cryptography>=41.0.1
dnspython>=2.3.0
bs4>=0.0.1
lxml>=4.9.3
coloredlogs>=15.0
